+++ src/transformer/descriptor/property/propertySignature.ts
@@ -1,7 +1,8 @@
 import * as ts from 'typescript';
 import { Scope } from '../../scope/scope';
 import { GetDescriptor } from '../descriptor';
-import { GetUndefinedDescriptor } from '../undefined/undefined';
+import { TransformerLogger } from '../../logger/transformerLogger';
+import { GetNullDescriptor } from '../null/null';
 import { PropertySignatureCache } from './cache';
 
 type PropertyNode = ts.PropertySignature | ts.PropertyDeclaration;
@@ -13,17 +14,12 @@ export function GetPropertyDescriptor(
   PropertySignatureCache.instance.set(node.name);
 
   if (node.type) {
-    if (node.questionToken) {
-      return GetUndefinedDescriptor();
-    }
-
     return GetDescriptor(node.type, scope);
   }
 
   if (!node.initializer) {
-    throw new Error(
-      `The transformer couldn't determine a property value for \`${node.getText()}' without a specified type nor an initializer value.`
-    );
+    TransformerLogger().typeOfPropertyNotFound(node);
+    return GetNullDescriptor();
   }
 
   return GetDescriptor(node.initializer, scope);